requirements.txt
Flask==2.0.3
PyYAML==6.0
mysql-connector-python==8.0.28
gunicorn==20.1.0
json-fix==0.1.3
